Update device scan limit scan new interface and special radix sort tests
* Update device scan limit scan new interface
* Update special radix sort tests
