+++ test/hipcub/CMakeLists.txt
@@ -91,6 +91,7 @@ endif()
 add_hipcub_test("hipcub.DeviceHistogram" test_hipcub_device_histogram.cpp)
 add_hipcub_test("hipcub.DeviceMergeSort" test_hipcub_device_merge_sort.cpp)
 add_hipcub_test("hipcub.DeviceRadixSort" test_hipcub_device_radix_sort.cpp)
+add_hipcub_test("hipcub.DeviceRadixSortSpecial" test_hipcub_device_radix_sort_floats.cpp)
 add_hipcub_test("hipcub.DeviceReduce" test_hipcub_device_reduce.cpp)
 add_hipcub_test("hipcub.DeviceRunLengthEncode" test_hipcub_device_run_length_encode.cpp)
 add_hipcub_test("hipcub.DeviceReduceByKey" test_hipcub_device_reduce_by_key.cpp)
